requirements.txt
streamlit
tavily-python
matplotlib
reportlab
python-dotenv
